requirements.txt
kivy==2.3.1
kivymd
peewee==3.17.1
google-generativeai==0.4.0
python-dotenv==1.0.1
pillow
